# openapi/jingxuan-v1.yaml
openapi: 3.1.0
info:
  title: Jingxuan API
  version: 1.0.0
security:
- BearerAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      operationId: post_auth_login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1LoginResponse'
          description: 登录成功
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Bad Request
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: 用户名或密码错误
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: 认证请求来源不受信任
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Not Found
        '405':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: 请求方法不受支持
        '406':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: 无法生成客户端可接受的响应格式
        '415':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: 请求媒体类型不受支持
        '422':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: 请求内容未通过业务校验
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Internal Server Error
      security: []
      summary: com/jingxuan/identityaccess/web/V1AuthController.java
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      operationId: post_auth_logout
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/V1RefreshRequest'
      responses:
        '200':
          description: OK
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Bad Request
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: 未提供有效身份凭证
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: 当前身份无权执行该操作
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Not Found
        '405':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: 请求方法不受支持
        '406':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: 无法生成客户端可接受的响应格式
        '415':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: 请求媒体类型不受支持
        '422':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: 请求内容未通过业务校验
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Internal Server Error
      summary: com/jingxuan/identityaccess/web/V1AuthController.java
      tags:
      - Auth
  /api/v1/auth/me:
    get:
      operationId: get_auth_me
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/V1UserInfo'
          description: OK
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Bad Request
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: 未提供有效身份凭证
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: 当前身份无权执行该操作
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Not Found
        '405':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: 请求方法不受支持
        '406':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: 无法生成客户端可接受的响应格式
        '422':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: 请求内容未通过业务校验
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Internal Server Error
      summary: com/jingxuan/identityaccess/web/V1AuthController.java
      tags:
      - Auth
  /api/v1/auth/refresh:
    post:
      operationId: post_auth_refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/V1RefreshRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1LoginResponse'
          description: 刷新成功
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Bad Request
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: 刷新令牌无效、过期或已被使用
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: 认证请求来源不受信任
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Not Found
        '405':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: 请求方法不受支持
        '406':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: 无法生成客户端可接受的响应格式
        '415':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: 请求媒体类型不受支持
        '422':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: 请求内容未通过业务校验
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Internal Server Error
      security: []
      summary: com/jingxuan/identityaccess/web/V1AuthController.java
      tags:
      - Auth
...
components:
  schemas:
    LoginRequest:
      description: 登录请求
      properties:
        password:
          description: 密码
          example: '123456'
          minLength: 1
          type: string
        rememberMe:
          description: 记住登录状态（7天）
          example: false
          type: boolean
        username:
          description: 用户名
          example: student001
          minLength: 1
          type: string
      required:
      - password
      - username
      type: object
    ProblemDetails:
      description: 统一 API 错误响应
      properties:
        code:
          description: 机器可读业务错误码
          type: string
        detail:
          description: 面向用户的错误详情
          type: string
        fieldErrors:
          additionalProperties:
            type: string
          description: 字段校验错误（键为字段名，值为错误消息）
          type: object
        instance:
          description: 请求实例 URI
          type: string
        requestId:
          description: 请求追踪 ID
          type: string
        status:
          description: HTTP 状态码
          format: int32
          type: integer
        title:
          description: 问题标题
          type: string
        type:
          description: 问题类型 URI
          type: string
      required:
      - code
      - detail
      - fieldErrors
      - instance
      - requestId
      - status
      - title
      - type
      type: object
...
    V1LoginResponse:
      properties:
        accessToken:
          type: string
        expiresIn:
          type: string
        refreshExpiresIn:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
        user:
          $ref: '#/components/schemas/V1UserInfo'
      type: object
...
    V1RefreshRequest:
      properties:
        refreshToken:
          minLength: 1
          type: string
      required:
      - refreshToken
      type: object
...
    V1UserInfo:
      properties:
        avatar:
          type: string
        classId:
          type: string
        className:
          type: string
        email:
          type: string
        firstLogin:
          type: boolean
        id:
          type: string
        phone:
          type: string
        realName:
          type: string
        roleCode:
          type: string
        roleId:
          format: int32
          type: integer
        roleName:
          type: string
        username:
          type: string
      type: object
...
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      description: '输入 Bearer Token，格式: Bearer {token}'
      scheme: bearer
      type: http